\name{young}
\alias{young}
\docType{data}
\title{Time series obtained by an fMRI experiement on the brain}
\description{
  Time series for each region of interest in the brain. These series are obtained by SPM preprocessing from young healthy individual. The length of the time series for each region of the brain is equal to 518.
}
\usage{data(young)}
\format{
  A data frame with 518 observations on the following 90 variables.
}
\details{

}
\source{
contact S. Achard (sa428@cam.ac.uk)
}
\references{
to appear
 }
\examples{
data(young)
## maybe str(brain) ; plot(brain) ...
}
\keyword{datasets}
